Created GUI for site_tester (rust)
Basic GUI with inputs
Still prints to console
Using egui

Projects Added:
    -

Projects Modified:
    - site_tester (rust)
