+++ site_tester/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2024"
 atty = "0.2.14"
 clap = "4.5.47"
 reqwest = { version = "0.12", default-features = false, features = ["blocking", "json", "rustls-tls"] }
+eframe = "0.27"
